// frontend/src/core/models/sugarfeed.ts
/**
 * Created by @ES Express Systems
 * User: Rafael Gutierrez Gaspar
 * Date: Mon Jan 25 2021 02:44:08 GMT-0400 (Bolivia Time)
 * Time: 2:44:8
 * Last User updated: Rafael Gutierrez Gaspar
 * Last date updated: Mon Jan 25 2021 02:44:08 GMT-0400 (Bolivia Time)
 * Last time updated: 2:44:8
 *
 * Caution: es-sections will be replaced by script execution
 */

//<es-section>

//</es-section>

export class Sugarfeed {

  //<es-section>
  
  id: string;
  
  
  
  
  
  name: string;
  
  related_module: string;
  
  link_url: string;
  
  link_type: string;
  
  
  description: string;
  
  
  
  date_entered: Date;
  
  date_modified: Date;
  
  
  
  modified_user_id: string;
  
  created_by: string;
  
  assigned_user_id: string;
  
  related_id: string;
  
  
  
  
  
  deleted: number;
  
  
  

  //</es-section>
}